+++ web/libs/ui/src/lib/select/types.ts
@@ -8,6 +8,12 @@ export type SelectOptionData<T = any> = {
   children?: SelectOptionData<T>[];
 };
 
+export enum SelectSize {
+  SMALL = "small",
+  MEDIUM = "medium",
+  LARGE = "large",
+}
+
 export type SelectOption<T> = string | number | SelectOptionData<T>;
 
 export type OptionProps = {
@@ -68,7 +74,7 @@ export type SelectProps<T, A extends SelectOption<T>[]> = {
   isInline?: boolean;
   isLoading?: boolean;
   dataTestid?: string;
-  size?: "small" | "medium" | "large";
+  size?: SelectSize | undefined;
   onSearch?: (value: string) => void;
   selectFirstIfEmpty?: boolean;
   renderSelected?: (selectedOptions?: A[number][], placeholder?: string) => React.ReactNode | string;
